python-script/requirements.txt
aiohttp>=3.8.0
paho-mqtt>=1.6.0
asyncio
requests>=2.28.0
dataclasses
pathlib
uuid
aiocoap>=0.4.6
pika>=1.3.0
websockets>=11.0
psutil>=5.8.0
matplotlib>=3.0.0
numpy>=1.21.0
pandas>=1.3.0 
